Don't require a returns section for `NoReturn`
When a function has a return type of `NoReturn` it means it will never
return, so the documentation doesn't really need to specify a return.

The option `--require-return-section-when-returning-none` was renamed to
`--require-return-section-when-returning-nothing` to highlight it is
more generic now, as it will also require a returns section for
`NoReturn`. The old option is kept as a deprecation error.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>