tox: use `deps = .` instead of `pip install`
If a user is using `tox-uv` (a plugin for tox that supports running the
various `pip` commands vi `uv`), then `pip install .` is seen as an
external command, which would require `allowlist_externals`. Instead,
install the local project by specifying it as a dep. Then, regular tox
and tox-uv both will install it as `pip install .` or `uv pip install
.`, respectively.