fix: check for parent ExceptHandler of Raise statements
Introduces a `walk.walk_dfs` for DFS traversal of a node. This allows us
to more easily keep track of direct parents of nested children, so that
we can find an ExceptHandler that may be containing any Raise statement.

Update tests for walk_dfs and some more complex edge cases for
try/except blocks.